[BUGFIX] Don't treat choices.js as ES6 module
The library choices.js was imported as a ES6 module which can fail as
the module is no ES6 module and thus has no exports. The issue was that
the code that would have been defined in exports is tried to get
executed, which fails with an error, which never popped up as the error
was not handled properly.

To solve this issue, the module is now loaded via `require()` which
solves the issue.

Resolves: DEV-836